requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
Pillow